#116 refactoring auth skeleton
